# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/Providers/getById/{id}": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Get single Providers",
                "description": "Returns Get single Providers",
                "operationId": "GetSingleProvider",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Providers id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Providers/create": {
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Insert new Provider",
                "description": "Returns Provider data",
                "operationId": "Insert Provider",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Provider"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/Providers/{id}/update": {
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Insert new Provider info",
                "description": "Returns Provider data",
                "operationId": "Update Provider",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Provider"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/Providers/{id}/delete": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Delete single Provider",
                "description": "Returns Deleted",
                "operationId": "DeleteSingleProvider",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/Providers/rate/{id}": {
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Insert Provider Rate",
                "description": "Returns Provider data",
                "operationId": "Rate Provider",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Provider"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/Providers/{id}/getRate": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Get single ProvidersRate",
                "description": "Returns Get single Providers Rate",
                "operationId": "GetSingleProviderRate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Providers id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Provider": {
                "title": "Provider",
                "description": "Provider model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "provider name",
                        "type": "string",
                        "example": "Sara"
                    },
                    "profession": {
                        "title": "profession",
                        "description": "provider job",
                        "type": "string",
                        "example": "dyer"
                    },
                    "skills": {
                        "title": "skills",
                        "description": "provider skills",
                        "type": "string",
                        "example": "carpentry"
                    },
                    "phone_number": {
                        "title": "phone_number",
                        "description": "provider phone",
                        "type": "string",
                        "example": "0770*******"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "provider picture",
                        "type": "string",
                        "example": "photo"
                    },
                    "models": {
                        "title": "models",
                        "description": "models of his work",
                        "type": "string",
                        "example": "photo"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Provider"
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}